fix: cluster migration v2 is merged to v1
